avoid implicit numerical conversion

+++ src/managers/curvemanager/curvemanagerquickaccessdelegate.cpp
@@ -72,9 +72,9 @@ void VisibilityArea::draw(QPainter& painter, const QModelIndex& index, const QRe
     }
   }();
 
-  icon = icon.scaledToWidth(rect.width(), Qt::SmoothTransformation);
+  icon = icon.scaledToWidth(static_cast<int>(rect.width()), Qt::SmoothTransformation);
   const auto y = static_cast<int>(rect.center().y() - icon.height() / 2.0);
-  painter.drawPixmap(rect.left(), y, icon);
+  painter.drawPixmap(static_cast<int>(rect.left()), y, icon);
 }
 
 void VisibilityArea::begin(const QModelIndex& index, QMouseEvent& event)
